refactor(utils): move `validate-id` to `/validators`
