+++ app/profile/[id]/top/genres/page.tsx
@@ -1,6 +1,6 @@
 import { redirect } from 'next/navigation'
 
-import { validateId } from '@app/utils/validate-id'
+import { validateId } from '@app/utils/validators'
 import {
   STATS_MEASUREMENT,
   STATS_PROVIDER,
