+++ app/utils/validators/index.ts
@@ -1,0 +1,1 @@
+export * from './validate-id'
